requirements.txt
azure-search-documents
azure-identity
python-dotenv
openai
azure-storage-blob  